separate WriteMode enum
